# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/subscriptions/current": {
      "get": {
        "description": "Get the current workspace subscription details including plan and status.",
        "operationId": "SubscriptionsController_getCurrentSubscription_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current subscription details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentSubscriptionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current subscription",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/usage": {
      "get": {
        "description": "Get the current usage for all features in the active workspace plan.",
        "operationId": "SubscriptionsController_getUsage_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usage statistics for all features",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get usage statistics",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/preview": {
      "get": {
        "description": "Preview the effects of changing to a different plan for the active workspace.",
        "operationId": "SubscriptionsController_previewPlanChange_v1",
        "parameters": [
          {
            "name": "planCode",
            "required": true,
            "in": "query",
            "description": "Target plan code (e.g., \"pro\", \"premium\", \"free\")",
            "schema": {
              "example": "pro",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan change preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview plan change",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/upgrade": {
      "post": {
        "description": "Upgrade the active workspace to a higher tier plan. The change takes effect immediately.",
        "operationId": "SubscriptionsController_upgradePlan_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan upgraded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpgradeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upgrade plan",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/downgrade": {
      "post": {
        "description": "Downgrade the active workspace to a lower tier plan. The change is scheduled for the end of the current billing period.",
        "operationId": "SubscriptionsController_downgradePlan_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Downgrade scheduled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DowngradeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Downgrade plan",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/cancel": {
      "post": {
        "description": "Cancel the active workspace subscription. It will downgrade to Free at the end of the billing period.",
        "operationId": "SubscriptionsController_cancelSubscription_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cancellation scheduled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel subscription",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/reactivate": {
      "post": {
        "description": "Reactivate a cancelled active workspace subscription before cancellation takes effect.",
        "operationId": "SubscriptionsController_reactivateSubscription_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Subscription reactivated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactivateResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivate subscription",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/scheduled": {
      "delete": {
        "description": "Cancel a scheduled downgrade or cancellation for the active workspace.",
        "operationId": "SubscriptionsController_cancelScheduledChange_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Scheduled change cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelScheduledChangeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel scheduled change",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/history": {
      "get": {
        "description": "Get a list of all plan changes for the active workspace.",
        "operationId": "SubscriptionsController_getPlanChangeHistory_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of records to return (default: 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan change history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanChangeHistoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get plan change history",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/workspaces/current": {
      "get": {
        "description": "Return the active workspace summary, membership context, and shared workspace settings. This route is intentionally workspace-scoped and does NOT expose personal user preferences.",
        "operationId": "WorkspacesController_getCurrentWorkspace_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current workspace summary and shared settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWorkspaceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current workspace",
        "tags": [
          "Workspaces"
        ]
      }
    },
    "/api/v1/workspaces/current/settings": {
      "get": {
        "description": "Return only shared workspace settings for the active workspace. Personal preferences remain user-scoped in UserPreference and are NOT read or mutated by this surface.",
        "operationId": "WorkspacesController_getCurrentWorkspaceSettings_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current shared workspace settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettingsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current workspace settings",
        "tags": [
          "Workspaces"
        ]
      },
      "patch": {
        "description": "Update shared workspace settings for the active workspace. Only workspace admins can mutate this surface. Personal user preferences must stay in user-scoped preference flows, not here.",
        "operationId": "WorkspacesController_updateCurrentWorkspaceSettings_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWorkspaceResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Only workspace admins can update shared workspace settings"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current workspace settings",
        "tags": [
          "Workspaces"
        ]
      }
    },
...
  },
  "info": {
    "title": "Facets API",
    "description": "Professional Finance Tracker SaaS API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SubscriptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Subscription ID"
          },
          "status": {
            "type": "string",
            "description": "Subscription status"
          },
          "currentPeriodStart": {
            "format": "date-time",
            "type": "string",
            "description": "Current billing period start"
          },
          "currentPeriodEnd": {
            "format": "date-time",
            "type": "string",
            "description": "Current billing period end (null for free plan)"
          },
          "trialStart": {
            "format": "date-time",
            "type": "string",
            "description": "Trial start date"
          },
          "trialEnd": {
            "format": "date-time",
            "type": "string",
            "description": "Trial end date"
          },
          "plan": {
            "description": "Current plan details",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "currentPeriodStart",
          "plan"
        ]
      },
      "CurrentSubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionDto"
          }
        },
        "required": [
          "subscription"
        ]
      },
      "FeatureUsageDto": {
        "type": "object",
        "properties": {
          "featureCode": {
            "type": "string",
            "description": "Feature code"
          },
          "current": {
            "type": "number",
            "description": "Current usage count"
          },
          "limit": {
            "type": "number",
            "description": "Maximum allowed (-1 for unlimited)"
          },
          "limitType": {
            "type": "string",
            "description": "Type of limit"
          },
          "featureType": {
            "type": "string",
            "description": "Feature type (RESOURCE or CONSUMABLE)"
          },
          "periodType": {
            "type": "string",
            "description": "Period type for consumable features"
          },
          "periodEnd": {
            "format": "date-time",
            "type": "string",
            "description": "Period end date for consumable features"
          },
          "usagePercentage": {
            "type": "number",
            "description": "Percentage of limit used (0-100+)"
          },
          "limitReached": {
            "type": "boolean",
            "description": "Whether the limit has been reached"
          }
        },
        "required": [
          "featureCode",
          "current",
          "limit",
          "limitType",
          "featureType",
          "usagePercentage",
          "limitReached"
        ]
      },
      "UsageResponseDto": {
        "type": "object",
        "properties": {
          "planCode": {
            "type": "string",
            "description": "Plan code"
          },
          "planName": {
            "type": "string",
            "description": "Plan name"
          },
          "features": {
            "description": "Usage details for each feature",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeatureUsageDto"
            }
          }
        },
        "required": [
          "planCode",
          "planName",
          "features"
        ]
      },
...
      "PlanChangePreviewDto": {
        "type": "object",
        "properties": {
          "currentPlan": {
            "description": "Current plan details",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanDto"
              }
            ]
          },
          "targetPlan": {
            "description": "Target plan details",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanDto"
              }
            ]
          },
          "changeType": {
            "type": "string",
            "description": "Type of change",
            "enum": [
              "UPGRADE",
              "DOWNGRADE"
            ]
          },
          "immediate": {
            "type": "boolean",
            "description": "Whether the change is applied immediately"
          },
          "effectiveAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the change will be applied (for downgrades)"
          },
          "prorationAmount": {
            "type": "number",
            "description": "Proration amount (positive = credit, negative = charge)",
            "example": 2.5
          },
          "overages": {
            "description": "Resources that exceed new plan limits",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceOverageDto"
            }
          },
          "hasOverages": {
            "type": "boolean",
            "description": "Whether there are any overages"
          },
          "gracePeriodEnd": {
            "format": "date-time",
            "type": "string",
            "description": "Grace period end date if there are overages"
          }
        },
        "required": [
          "currentPlan",
          "targetPlan",
          "changeType",
          "immediate",
          "overages",
          "hasOverages"
        ]
      },
      "PreviewResponseDto": {
        "type": "object",
        "properties": {
          "preview": {
            "$ref": "#/components/schemas/PlanChangePreviewDto"
          }
        },
        "required": [
          "preview"
        ]
      },
      "ChangePlanDto": {
        "type": "object",
        "properties": {
          "planCode": {
            "type": "string",
            "description": "Target plan code to change to",
            "example": "pro"
          }
        },
        "required": [
          "planCode"
        ]
      },
      "UpgradeResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Successfully upgraded to Pro plan"
          },
          "subscription": {
            "description": "Updated subscription details",
            "allOf": [
              {
                "$ref": "#/components/schemas/SubscriptionDto"
              }
            ]
          },
          "prorationAmount": {
            "type": "number",
            "description": "Proration amount applied"
          }
        },
        "required": [
          "message",
          "subscription"
        ]
      },
      "DowngradeResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Downgrade to Free plan scheduled for end of billing period"
          },
          "subscription": {
            "description": "Current subscription (plan not changed yet)",
            "allOf": [
              {
                "$ref": "#/components/schemas/SubscriptionDto"
              }
            ]
          },
          "scheduledFor": {
            "format": "date-time",
            "type": "string",
            "description": "When the downgrade will be applied"
          },
          "targetPlanCode": {
            "type": "string",
            "description": "Target plan code"
          },
          "overages": {
            "description": "Resources that will exceed new plan limits",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceOverageDto"
            }
          },
          "gracePeriodEnd": {
            "format": "date-time",
            "type": "string",
            "description": "Grace period end date if there are overages"
          }
        },
        "required": [
          "message",
          "subscription",
          "scheduledFor",
          "targetPlanCode",
          "overages"
        ]
      },
      "CancelSubscriptionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Reason for cancellation",
            "example": "Too expensive",
            "maxLength": 500
          }
        }
      },
      "CancelResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Subscription cancellation scheduled"
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the subscription will be cancelled"
          },
          "effectiveAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the subscription will effectively end"
          }
        },
        "required": [
          "message",
          "cancelledAt",
          "effectiveAt"
        ]
      },
      "ReactivateResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Subscription reactivated successfully"
          },
          "subscription": {
            "description": "Updated subscription details",
            "allOf": [
              {
                "$ref": "#/components/schemas/SubscriptionDto"
              }
            ]
          }
        },
        "required": [
          "message",
          "subscription"
        ]
      },
      "CancelScheduledChangeResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Scheduled change cancelled successfully"
          },
          "subscription": {
            "description": "Updated subscription details",
            "allOf": [
              {
                "$ref": "#/components/schemas/SubscriptionDto"
              }
            ]
          }
        },
        "required": [
          "message",
          "subscription"
        ]
      },
...
      "PlanChangeHistoryResponseDto": {
        "type": "object",
        "properties": {
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanChangeLogDto"
            }
          }
        },
        "required": [
          "history"
        ]
      },
...
      "CurrentWorkspaceResponseDto": {
        "type": "object",
        "properties": {
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceSummaryDto"
          },
          "membership": {
            "$ref": "#/components/schemas/WorkspaceMembershipSummaryDto"
          },
          "settings": {
            "$ref": "#/components/schemas/WorkspaceSettingsDto"
          }
        },
        "required": [
          "workspace",
          "membership",
          "settings"
        ]
      },
      "WorkspaceSettingsResponseDto": {
        "type": "object",
        "properties": {
          "settings": {
            "$ref": "#/components/schemas/WorkspaceSettingsDto"
          }
        },
        "required": [
          "settings"
        ]
      },
      "UpdateWorkspaceSettingsDto": {
        "type": "object",
        "properties": {
          "baseCurrencyCode": {
            "type": "string",
            "description": "Workspace base currency code shared by all members",
            "example": "ARS"
          },
          "baseLanguage": {
            "type": "string",
            "description": "Workspace base language shared by all members",
            "example": "es"
          },
          "dateFormat": {
            "type": "string",
            "description": "Workspace-wide default date format",
            "example": "YYYY-MM-DD"
          },
          "monthStartDay": {
            "type": "number",
            "description": "Day of month that starts the financial month",
            "example": 5
          },
          "weekStartDay": {
            "type": "number",
            "description": "Day of week that starts the calendar week (1-7)",
            "example": 1
          },
          "timezone": {
            "type": "string",
            "description": "Workspace default timezone",
            "example": "America/Argentina/Buenos_Aires"
          },
          "locale": {
            "type": "string",
            "description": "Workspace default locale",
            "example": "es-AR"
          },
          "displayLabel": {
            "type": "string",
            "description": "Representative label shown for the workspace across the product",
            "example": "Casa"
          },
          "workspaceType": {
            "type": "string",
            "description": "Workspace type used for shared workspace semantics",
            "enum": [
              "PERSONAL",
              "FAMILY",
              "GROUP"
            ],
            "example": "FAMILY"
          }
        }
      },
...
    }
  }
}